- use +/- to change tess level

+++ Dx12Tessellation/Dx12Tessellation.h
@@ -25,6 +25,6 @@ protected:
 	}
 
 private:
-	UINT m_tesstriTessLevel;
+	FLOAT m_tesstriTessLevel;
 
 };
